ruff fixes

+++ pwndck/db_size.py
@@ -26,7 +26,7 @@ def estimate_db(samples: int = 10):
 
     lengths = [get_line_count(x) for x in hashes]
 
-    scale = int(2**20)
+    scale = 2**20
 
     mean = scale * statistics.mean(lengths)
     stdev = scale * statistics.stdev(lengths)
@@ -38,7 +38,7 @@ def sig_figs(x: float, precision: int) -> float:
     x = float(x)
     precision = int(precision)
 
-    return round(x, -int(floor(log10(abs(x)))) + (precision - 1))
+    return round(x, -floor(log10(abs(x))) + (precision - 1))
 
 
 def fmt_num(num: float, precision: int) -> str:
